+++ latte/models/probabilistic_pca/evaluation.py
@@ -6,7 +6,7 @@ import torch
 from latte.dataset import utils as dsutils
 from latte.models.probabilistic_pca import probabilistic_pca
 from latte.utils import evaluation
-from latte.metrics import mig
+from latte.evaluation import metrics
 
 
 class OrientationMetrics:
@@ -175,9 +175,11 @@ def evaluate_full_result(
         sigma (float): The true observable data noise standard deviation.
         sigma_hat (float): The estimate of the observable data noise standard deviation.
         stopped_epoch (int): Number of epochs required to orient the model.
-        n_runs (int, optional): How many random matrices to use to calculate the mean distance to a permutation matrix
-                                of a random matrix orthogonal matrix.
-                                Defaults to 50.
+        train_loss (float): The recorded final train loss of the orientation.
+        loss (float): The recorded final test loss of the orientation.
+        nruns (int, optional): How many random matrices to use to calculate the mean distance to a permutation matrix
+                               of a random matrix orthogonal matrix.
+                               Defaults to 50.
 
     Returns:
         pd.DataFrame: A data frame containing the values of the metrics grouped in the OrientationMetrics class.
@@ -196,7 +198,7 @@ def evaluate_full_result(
     evaluation_results.append(
         (
             OrientationMetrics.DISTANCE_TO_PERMUTATION_MATRIX,
-            round(evaluation.distance_to_permutation_matrix(torch.from_numpy(R)), 3),
+            round(metrics.distance_to_permutation_matrix(torch.from_numpy(R)), 3),
         )
     )
     evaluation_results.append(
@@ -206,9 +208,7 @@ def evaluate_full_result(
                 float(
                     np.mean(
                         [
-                            evaluation.distance_to_permutation_matrix(
-                                geoopt.Stiefel().random(max(R.shape), min(R.shape))
-                            )
+                            metrics.distance_to_permutation_matrix(geoopt.Stiefel().random(max(R.shape), min(R.shape)))
                             for _ in range(nruns)
                         ]
                     )
@@ -243,8 +243,8 @@ def evaluate_full_result(
     Z_pca_oriented = probabilistic_pca.get_latent_representations(X, A_hat_oriented, mu_hat, sigma_hat)
 
     # Calculate the MIG score of the original and the oriented representations
-    evaluation_results.append((OrientationMetrics.ORIGINAL_MIG, mig.mig(Z_pca_original, Z)))
-    evaluation_results.append((OrientationMetrics.ORIENTED_MIG, mig.mig(Z_pca_oriented, Z)))
+    evaluation_results.append((OrientationMetrics.ORIGINAL_MIG, metrics.mig(Z_pca_original, Z)))
+    evaluation_results.append((OrientationMetrics.ORIENTED_MIG, metrics.mig(Z_pca_oriented, Z)))
 
     # Calculate the mean latent error from the true observations
     r_error_oriented = np.linalg.norm(Z_pca_oriented - Z, axis=1).mean()
