+++ latte/utils/evaluation.py
@@ -9,7 +9,6 @@ from itertools import product
 import numpy as np
 import pandas as pd
 import torch
-from scipy import optimize
 
 from sklearn import feature_selection
 
@@ -88,38 +87,3 @@ def axis_factor_mutual_information(Z: torch.Tensor, Y: torch.Tensor, factor_name
         # we simulate 1 feature, we extract the first one
 
     return pd.DataFrame(mis)
-
-
-def distance_to_permutation_matrix(R: torch.Tensor) -> float:
-    """
-    Calculates the Frobenius norm of the closest *signed* permutation matrix as measured by the Fribenius norm.
-    The closest signed permutation matrix is computed using the linear sum assignment algorithm with the cost
-    matrix defined by the contributions of the permutation matrix entries to the Frobenius norm.
-    *Importantly*, this assumes that the matrix R is *orthogonal*.
-    Implemented with the help of
-    https://math.stackexchange.com/questions/2951811/nearest-signed-permutation-matrix-to-a-given-matrix-a.
-
-    Args:
-        R (np.ndarray): The projection matrix to evaluate.
-
-    Returns:
-        float: The Frobenius norm to the closest signed permutation matrix.
-    """
-    # assert mutils.is_orthonormal(R, atol=5e-2)
-
-    # The cost matrix for the Hungarian algorithm
-    # In the general case, the first term would include the *squared L2 row norms*, however, in the orthogonal case,
-    # they are zero
-    # See the link above to see what the elements of this matrix corresponds to
-    # C = 1 - R**2 + (torch.abs(R) - 1) ** 2
-
-    # The general case
-    C = torch.linalg.norm(R, dim=1).reshape(-1, 1) ** 2 - R**2 + (R.abs() - 1) ** 2
-
-    M = optimize.linear_sum_assignment(C)  # M includes the rows and columns of where ones should go
-
-    P = torch.zeros(R.shape)
-    P[M] = 1  # Put ones into the entries where they should be
-    P *= torch.sign(R)  # Multiply with the sign to take into account the absolute values from the calculation of `C`
-
-    return torch.linalg.norm(R - P).item()
